Fix potential read out of bounds
The vxi11_open_device function keeps a list of known clients which
includes the address.  This is a string which can be an IP address
(of fixed length) but could also be host names.  Change the structure
to use a zero-length array for the address and allocate memory
according to the address string's length.