Add Fedora .spec file
With this .spec file the package is compiled and packaged in four
different RPM files (plus the debug information).  The library,
header, utilities, and the Python code can be installed separately.
The library is a dependency for the other packages, though.

Because Fedora (and upstream) for some time has deprecated Python 2
the code is only packaged for Python 3.